% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brexit.R
\docType{data}
\name{brexit}
\alias{brexit}
\title{This is the first data to be included in my package}
\description{
it contains news articles on brexits
}
\author{
SimiKafaru \email{kafarusimileoluwa@gmail.com}
}
\keyword{data}
