% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Counter.R
\name{counter}
\alias{counter}
\title{Get the number of times a vector of words occurs}
\usage{
counter(word_vec, words)
}
\arguments{
\item{word_vec}{is the text column}

\item{words}{this is a vector of words you want to retrieve their frequency counts}
}
\description{
This function retrieves the number of times each word in a vector occurs.
It returns a dataframe containing the word and the corresponding counts
}
