% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doge.R
\docType{data}
\name{doge}
\alias{doge}
\title{This is the first data to be included in my package}
\description{
it contains tweets on doge coin collected using twitter API
}
\author{
SimiKafaru \email{kafarusimileoluwa@gmail.com}
}
\keyword{data}
