% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrc.R
\name{nrc}
\alias{nrc}
\title{This data was saved
NRC word-emotion association lexicon}
\source{
\url{http://saifmohammad.com/WebPages/lexicons.html}
}
\value{
A tibble with 13,901 rows and 4 variables:
\describe{
\item{word}{An English word}
\item{sentiment}{Indicator for sentiment or emotion: "negative",
"positive", "anger", "anticipation", "disgust", "fear", "joy", "sadness",
"surprise", or "trust"}
}
}
\description{
The dataset is saved from the textdata\url{https://github.com/EmilHvitfeldt/textdata/blob/master/R/lexicon_nrc.R} for easier access
}
