% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNumPunct.R
\name{removeNumPunct}
\alias{removeNumPunct}
\title{Easily remove Punctuation from Text}
\usage{
removeNumPunct(x)
}
\arguments{
\item{x}{is the text column you want the punctuation and texts removed from}
}
\description{
This function will help you remove punctuation and numbers from your text easily
}
\examples{
{
removeNumPunct("is this your number? 01234")

}

}
