% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeURL.R
\name{removeURL}
\alias{removeURL}
\title{A function to help you remove URLs from text}
\usage{
removeURL(x)
}
\arguments{
\item{x}{is the text value you want to extract the texts from}
}
\value{
a character vector.
}
\description{
This function helps remove URLs from text, particularly designed for tweets
}
