% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeURL2.R
\name{removeURL2}
\alias{removeURL2}
\title{A function to help you remove URLs from text}
\usage{
removeURL2(x)
}
\arguments{
\item{x}{is the text value you want to extract the texts from}
}
\description{
This function helps remove URLs from text
}
