% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentimentAnalyzer.R
\name{sentimentAnalyzer}
\alias{sentimentAnalyzer}
\title{Get the overall weight of emotions conveyed in a tweet}
\usage{
sentimentAnalyzer(word_vec, details)
}
\arguments{
\item{word_vec}{This is the corpus you want to extract the sentiments from}

\item{details}{(A TRUE/FALSE value): If TRUE you get a more robust distribution of these emotions. FALSE is summarised as Positive or Negative}
}
\description{
This function will help you extract the weight of emotions conveyed in a tweet
}
\examples{

sentimentAnalyzer(doge$text, details = T)

}
