% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_words.R
\docType{data}
\name{stop_words}
\alias{stop_words}
\title{Saved stop_word dataframe from tidytext}
\description{
it contains stop_words from tidytext package.
It is saved for easier acces
}
\author{
tidytext
}
\keyword{data}
