% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_bigrams.R
\name{top_bigrams}
\alias{top_bigrams}
\title{Get the top bigrams from text
Get the top n bigrams from vector of text}
\usage{
top_bigrams(word_vec, remove_these, bigram_size)
}
\arguments{
\item{word_vec}{This is the corpus you want to extract the sentiments from}

\item{remove_these}{This is a vector of characters you want cleaned out of the text}

\item{bigram_size}{This is the Top N number of rows to be retrieved as an integer value}
}
\value{
a data frame object.
}
\description{
This function is used to get the top N bigrams from a corpus. It will retrieve the most occurring two combinations based on frequency
}
\examples{

{
top_bigrams(brexit[, c("content")], remove_these = c("rt"), bigram_size = 20)
}

}
