% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_words.R
\name{top_words}
\alias{top_words}
\title{Get the top n words from vector of text}
\usage{
top_words(word_vec, remove_these, size)
}
\arguments{
\item{word_vec}{is the variable containing text to be analyzed}

\item{remove_these}{is a vector of characters you want cleaned out of the tex}

\item{size}{is the N number of rows to be retrieved as an integer value}
}
\description{
This function is used to get the top N words from a vactor of text. It will retrieve the most occuring words based on frequency
}
\examples{

{
top_words(df$content, remove_these = c("news","uk"), size = 10)
}

}
