% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_words_Retriever.R
\name{top_words_Retriever}
\alias{top_words_Retriever}
\title{Get the top words based on a key search word}
\usage{
top_words_Retriever(word_vec, word_ret, remove_these, size)
}
\arguments{
\item{word_vec}{This is the corpus you want to extract the sentiments from}

\item{word_ret}{is the key word you want searched}

\item{remove_these}{is a vector of characters you want cleaned out of the tex}

\item{size}{is the N number of rows to be retrieved as an integer value}
}
\description{
This function helps to search for the top n words but only based texts or rows
containing a key word. It is particularly useful when you want to search the top n words revolving
around a certain keyword
}
\examples{

{
top_words_Retriever(brexit$content, word_ret = "brexit", remove_these = c("news","uk"), size = 10)
}

}
