% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{users}
\alias{users}
\title{Extract Usernames and tagged handles from tweets}
\usage{
users(x, ...)
}
\arguments{
\item{x}{This is the corpus you want to extract the mentions from}
}
\description{
The function will any tagged handles from text
}
\examples{
{
users("Come See this @simmie_kafaru")
}


}
