% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\name{counter}
\alias{counter}
\title{Get the number of times a vector of words occurs}
\usage{
counter(word_vec, words)
}
\arguments{
\item{word_vec}{This is the corpus you want to the word frequency extracted from}

\item{words}{This is a vector of words you want to retrieve their frequency counts}
}
\value{
a data frame object. A data frame object of strings and their corresponding count
}
\description{
This function retrieves the number of times each word in a corpus occurs.
It returns a dataframe containing the word and the corresponding counts
}
