% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_Sentiments.R
\name{top_Sentiments}
\alias{top_Sentiments}
\title{Get the top 10 negative and positive words}
\usage{
top_Sentiments(word_vec, plot)
}
\arguments{
\item{word_vec}{This is the corpus you want to extract the sentiments from}

\item{plot}{(TRUE/FALSE) TRUE means you want to return a plot which you can further customize. FALSE means a dataframe will be returned}
}
\value{
a data frame object if plot is FALSE. a ggplot object if plot = TRUE
}
\description{
This function returns the top 10 positive and negative words expressed in a text.
By defaults a data frame of words classified as positive or negative based on weights.
}
\examples{
top_Sentiments(doge$text, plot = TRUE)

}
