% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_words.R
\name{top_words}
\alias{top_words}
\title{Get the top n words from vector of text}
\usage{
top_words(word_vec, remove_these, size)
}
\arguments{
\item{word_vec}{This is the corpus you want to extract the sentiments from}

\item{remove_these}{This is a vector of characters you want cleaned out of the text}

\item{size}{This is the Top N number of rows to be retrieved as an integer value}
}
\value{
a data frame object.
}
\description{
This function is used to get the top N words from a corpus. It will retrieve the most occurring words based on frequency
}
\examples{

{
top_words(brexit$content, remove_these = c("news","uk"), size = 10)
}

}
